[Fix] Fix timer position and size on iPhone 6 [Fix] Adding support for iPhone 6 [Fix] Replacing "%d" with "%ld" for better 64 bits support

+++ ImportFromWebsiteViewController_Phone.m
@@ -240,7 +240,7 @@
 #if TARGET_IPHONE_SIMULATOR
 	NSInteger currentCount = [Countdown allCountdowns].count;
 	NSInteger importedCount = selectedCountdowns.count;
-	NSDebugLog(@"%i from current countdowns + %i imported countdowns", currentCount, importedCount);
+	NSDebugLog(@"%li from current countdowns + %ld imported countdowns", (long)currentCount, (long)importedCount);
 #endif
 	
 	[Countdown addCountdowns:selectedCountdowns];
